--- .idea/libraries/Maven__javax_inject_javax_inject_1.xml
@@ -1,13 +1,0 @@
-<component name="libraryTable">
-  <library name="Maven: javax.inject:javax.inject:1">
-    <CLASSES>
-      <root url="jar://$MAVEN_REPOSITORY$/javax/inject/javax.inject/1/javax.inject-1.jar!/" />
-    </CLASSES>
-    <JAVADOC>
-      <root url="jar://$MAVEN_REPOSITORY$/javax/inject/javax.inject/1/javax.inject-1-javadoc.jar!/" />
-    </JAVADOC>
-    <SOURCES>
-      <root url="jar://$MAVEN_REPOSITORY$/javax/inject/javax.inject/1/javax.inject-1-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
