--- .idea/libraries/Maven__org_springframework_social_spring_social_twitter_1_1_2_RELEASE.xml
@@ -1,13 +1,0 @@
-<component name="libraryTable">
-  <library name="Maven: org.springframework.social:spring-social-twitter:1.1.2.RELEASE">
-    <CLASSES>
-      <root url="jar://$MAVEN_REPOSITORY$/org/springframework/social/spring-social-twitter/1.1.2.RELEASE/spring-social-twitter-1.1.2.RELEASE.jar!/" />
-    </CLASSES>
-    <JAVADOC>
-      <root url="jar://$MAVEN_REPOSITORY$/org/springframework/social/spring-social-twitter/1.1.2.RELEASE/spring-social-twitter-1.1.2.RELEASE-javadoc.jar!/" />
-    </JAVADOC>
-    <SOURCES>
-      <root url="jar://$MAVEN_REPOSITORY$/org/springframework/social/spring-social-twitter/1.1.2.RELEASE/spring-social-twitter-1.1.2.RELEASE-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
