--- src/test/java/org/eaa690/aerie/model/BaseEntityTests.java
@@ -1,75 +1,0 @@
-/*
- *  Copyright (C) 2021 Gwinnett County Experimental Aircraft Association
- *
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *  you may not use this file except in compliance with the License.
- *  You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
- */
-
-package org.eaa690.aerie.model;
-
-import java.util.Date;
-import org.junit.Assert;
-import org.junit.Before;
-import org.junit.Test;
-
-import org.eaa690.aerie.exception.ResourceNotFoundException;
-
-/**
- * BaseEntityTests.
- */
-public class BaseEntityTests {
-
-    /**
-     * TEXT.
-     */
-    public static final Date DATE = new Date();
-
-    /**
-     * BaseEntity.
-     */
-    private BaseEntity baseEntity;
-
-    /**
-     * Test setup.
-     *
-     * @throws ResourceNotFoundException when a test error occurs
-     */
-    @Before
-    public void before() throws ResourceNotFoundException {
-        baseEntity = new BaseEntity();
-    }
-
-    /**
-     * Test setting created at.
-     *
-     * @throws Exception when a test error occurs
-     */
-    @Test
-    public void setCreatedAt() throws Exception {
-        baseEntity.setCreatedAt(DATE);
-
-        Assert.assertNotNull(baseEntity.getCreatedAt());
-    }
-
-    /**
-     * Test setting updated at.
-     *
-     * @throws Exception when a test error occurs
-     */
-    @Test
-    public void setUpdatedAt() throws Exception {
-        baseEntity.setUpdatedAt(DATE);
-
-        Assert.assertNotNull(baseEntity.getUpdatedAt());
-    }
-
-}
