--- src/test/java/org/eaa690/aerie/model/wx/ElevationTests.java
@@ -1,85 +1,0 @@
-/*
- *  Copyright (C) 2021 Gwinnett County Experimental Aircraft Association
- *
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *  you may not use this file except in compliance with the License.
- *  You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
- */
-
-package org.eaa690.aerie.model.wx;
-
-import org.eaa690.aerie.exception.ResourceNotFoundException;
-import org.junit.Assert;
-import org.junit.Before;
-import org.junit.Test;
-
-/**
- * ElevationTests.
- */
-public class ElevationTests {
-
-    /**
-     * TEXT.
-     */
-    public static final String TEXT = "ABC123";
-
-    /**
-     * Elevation.
-     */
-    private Elevation elevation;
-
-    /**
-     * Test setup.
-     *
-     * @throws ResourceNotFoundException when a test error occurs
-     */
-    @Before
-    public void before() throws ResourceNotFoundException {
-        elevation = new Elevation();
-    }
-
-    /**
-     * Test setting feet.
-     *
-     * @throws Exception when a test error occurs
-     */
-    @Test
-    public void setFeet() throws Exception {
-        elevation.setFeet(TEXT);
-
-        Assert.assertNotNull(elevation.getFeet());
-    }
-
-    /**
-     * Test setting meters.
-     *
-     * @throws Exception when a test error occurs
-     */
-    @Test
-    public void setMeters() throws Exception {
-        elevation.setMeters(TEXT);
-
-        Assert.assertNotNull(elevation.getMeters());
-    }
-
-    /**
-     * Test setting method.
-     *
-     * @throws Exception when a test error occurs
-     */
-    @Test
-    public void setMethod() throws Exception {
-        elevation.setMethod(TEXT);
-
-        Assert.assertNotNull(elevation.getMethod());
-    }
-
-}
