--- src/test/java/org/eaa690/aerie/steps/BaseSteps.java
@@ -1,56 +1,0 @@
-/*
- *  Copyright (C) 2021 Gwinnett County Experimental Aircraft Association
- *
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *  you may not use this file except in compliance with the License.
- *  You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
- */
-
-package org.eaa690.aerie.steps;
-
-import java.io.IOException;
-import java.util.Properties;
-import io.restassured.response.Response;
-import org.eaa690.aerie.AerieApplicationTests;
-import org.eaa690.aerie.TestContext;
-
-/**
- * Base class for Cucumber test implementation classes.
- */
-public class BaseSteps extends AerieApplicationTests {
-
-    private static final boolean PRINT = false;
-
-    private static Properties properties = new Properties();
-
-    /**
-     * TestContext.
-     */
-    protected TestContext testContext = TestContext.getInstance();
-
-    public BaseSteps() {
-        try {
-            properties.load(getClass().getResourceAsStream("/application.properties"));
-        } catch (IOException ioe) {
-            System.out.println("ERROR loading properties: " + ioe.getMessage());
-        }
-    }
-
-    protected String getAerieBaseUrl() {
-        return (String) properties.get("aerie.base-url");
-    }
-
-    protected void printResponse(Response response) {
-        if (PRINT) {
-            System.out.println("response: " + response.prettyPrint());
-        }
-    }
-}
