+++ src/test/resources/application.properties
@@ -14,4 +14,4 @@
 #  limitations under the License.
 #
 
-aerie.base-url="http://localhost:8080/"
+aerie.base-url=http://localhost:8080/
